@@ -1,6 +1,6 @@
 {
   "name": "valentinusz-github-io",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
@@ -14,24 +14,23 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.3",
-    "@docusaurus/preset-classic": "^2.4.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "@mdx-js/react": "^1.6.22",
-    "reactflow": "^11.9.4",
+    "@docusaurus/core": "^3.0.0-rc.1",
+    "@docusaurus/preset-classic": "^3.0.0-rc.1",
+    "@mdx-js/react": "^3.0.0",
     "@reactflow/background": "^11.3.4",
     "@reactflow/controls": "^11.2.4",
-    "clsx": "^1.2.1",
+    "clsx": "^2.0.0",
     "docusaurus-plugin-sass": "^0.2.5",
-    "hast-util-is-element": "^1.1.0",
-    "prism-react-renderer": "^1.3.5",
-    "rehype-katex": "^5.0.0",
-    "remark-math": "^3.0.1"
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "reactflow": "^11.9.4",
+    "rehype-katex": "^7.0.0",
+    "remark-math": "^6.0.0"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "^2.4.3",
-    "sass": "^1.69.4",
+    "sass": "^1.69.5",
     "sass-loader": "^13.3.2"
   },
   "browserslist": {
@@ -47,6 +46,6 @@
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
