@@ -1,17 +1,5 @@
-/**
- * Creating a sidebar enables you to:
- - create an ordered group of notes
- - render a sidebar for each doc of that group
- - provide next/previous navigation
-
- The sidebars can be generated from the filesystem, or explicitly defined here.
-
- Create as many sidebars as you want.
- */
-
+// noinspection JSUnusedGlobalSymbols
 /** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
-  tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+export default {
+  notesSidebar: [{type: 'autogenerated', dirName: '.'}],
 };
-
-module.exports = sidebars;
