@@ -1,7 +1,5 @@
-import React from "react";
 import styles from './styles.module.scss';
-import {Handle} from "reactflow";
-import {Position} from "reactflow";
+import {Handle, Position} from "reactflow";
 import DatabaseTable from "../DatabaseTable/DatabaseTable";
 
 /**
@@ -12,10 +10,10 @@ import DatabaseTable from "../DatabaseTable/DatabaseTable";
 export default function DatabaseTableNode({data: {name, columns}}) {
     return (
         <div className={styles.container}>
-            <Handle id='top' type="source" position={Position.Top} />
-            <Handle id='bottom' type="source" position={Position.Bottom} />
-            <Handle id='left' type="source" position={Position.Left} />
-            <Handle id='right' type="source" position={Position.Right} />
+            <Handle id='top' type='source' position={Position.Top} />
+            <Handle id='bottom' type='source' position={Position.Bottom} />
+            <Handle id='left' type='source' position={Position.Left} />
+            <Handle id='right' type='source' position={Position.Right} />
             <h3>{name}</h3>
             <DatabaseTable className={styles.node} columns={columns}/>
         </div>
