@@ -1,5 +1,4 @@
-import React, {useRef} from "react";
-import { useCallback, useState } from 'react';
+import { useRef, useCallback, useState } from 'react';
 import ReactFlow, {addEdge, applyEdgeChanges, applyNodeChanges, Background, Controls} from 'reactflow';
 import 'reactflow/dist/style.css';
 
@@ -7,8 +6,6 @@ import DatabaseTableNode from "../DatabaseTableNode/DatabaseTableNode";
 import {ControlButton} from "@reactflow/controls";
 import OpenInFullIcon from "../../../static/img/open_in_full.svg"
 
-// we define the nodeTypes outside the component to prevent re-renderings
-// you could also use useMemo inside the component
 const nodeTypes = { databaseTable: DatabaseTableNode };
 
 export default function SystemCatalogsFlow({initialNodes = [], initialEdges = []}) {
@@ -26,10 +23,12 @@ export default function SystemCatalogsFlow({initialNodes = [], initialEdges = []
         (changes) => setNodes((nds) => applyNodeChanges(changes, nds)),
         [setNodes]
     );
+
     const onEdgesChange = useCallback(
         (changes) => setEdges((eds) => applyEdgeChanges(changes, eds)),
         [setEdges]
     );
+
     const onConnect = useCallback(
         (connection) => setEdges((eds) => addEdge(connection, eds)),
         [setEdges]
@@ -44,26 +43,20 @@ export default function SystemCatalogsFlow({initialNodes = [], initialEdges = []
     }
 
     return (
-        <>
-            <ReactFlow
-                ref={flowElement}
-                nodes={nodes}
-                edges={edges}
-                onNodesChange={onNodesChange}
-                onEdgesChange={onEdgesChange}
-                onConnect={onConnect}
-                nodeTypes={nodeTypes}
-                connectionMode="loose"
-                nodesConnectable={false}
-                fitView
-            >
-                <Background color="#ccc" variant='dots' />
-                <Controls >
-                    <ControlButton onClick={toggleFullScreen} title='fullscreen'>
-                        <OpenInFullIcon/>
-                    </ControlButton>
-                </Controls>
-            </ReactFlow>
-        </>
+        <ReactFlow
+            ref={flowElement}
+            nodes={nodes} onNodesChange={onNodesChange}
+            edges={edges} onEdgesChange={onEdgesChange}
+            nodeTypes={nodeTypes}
+            onConnect={onConnect} connectionMode='loose' nodesConnectable={false}
+            fitView
+        >
+            <Background color='#ccc' variant='dots'/>
+            <Controls>
+                <ControlButton onClick={toggleFullScreen} title='fullscreen'>
+                    <OpenInFullIcon/>
+                </ControlButton>
+            </Controls>
+        </ReactFlow>
     );
 }
